+++ README.md
@@ -1,1 +1,4 @@
-![demo](https://s3.meituan.net/v1/mss_814dc1610cda4b2e8febd6ea2c809db5/apps-open/d393e7d1-3c0f-4fdb-9e90-b2a6bb07028a_1493714768227?filename=1493714768045WU_FILE_0.png)
+最大态
+![image.png](https://s3.meituan.net/v1/mss_814dc1610cda4b2e8febd6ea2c809db5/apps-open/f5feba03-1ecd-4fe1-8990-486c2f868616_1495808280036?filename=image.png)
+中间态
+![image.png](https://s3.meituan.net/v1/mss_814dc1610cda4b2e8febd6ea2c809db5/apps-open/f5feba03-1ecd-4fe1-8990-486c2f868616_1495808280036?filename=image.png)
