+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "classnames": "^2.2.5",
     "hotkeys-js": "^2.0.1",
+    "interactjs": "^1.2.8",
     "mobx": "^3.1.9",
     "mobx-react": "^4.1.8",
     "omit.js": "^0.1.0",
@@ -68,10 +69,10 @@
     "node": ">= 5.0.0",
     "npm": ">= 3.0.0"
   },
-    "files": [
-        "dist",
-        "src",
-        "index.js",
-        "README.md"
-    ]
+  "files": [
+    "dist",
+    "src",
+    "index.js",
+    "README.md"
+  ]
 }
